simulator/requirements.txt
# Core dependencies for Haystack Building Data Simulator
psycopg2-binary>=2.9.0
pandas>=1.5.0
numpy>=1.24.0
PyYAML>=6.0
rich>=13.0.0
SQLAlchemy>=2.0.0

# Service and scheduling dependencies
APScheduler>=3.10.0

# Optional dependencies for extended functionality
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0